Fix #2 #3 support revision at non-current branch
If parsing `@foo` results in valid revision, but git-log can't find an
entry in current branch, we can still view the diff, only shortage is
that history context (i.e., `j` `k` shortcuts) are not available.

There is conflict that we use `@` for shortcut, and git takes `@` as a
shorthand for `HEAD`. To reduce complexity, we try normalize `@` to
`HEAD` before parse.