Do extra diffupdate after new window diff ready
Problem: when gitdiffall open side-by-side diff for specific revision,
it is hard to do some action on diff complete.

With autocmd `OptionSet diff`, it only works on 2nd opened window
because we do:

1) new window A, diffthis
2) vertical split window B, diffthis

where step 1) won't trigger the autocmd (reproduced Vim 8.1.1835).

Workaround is to force a `diffupdate` on window A, and use `DiffUpdated`
autocmd to watch it.